backend/requirements.txt
dspy-ai==2.5.6
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2